test: enhance data race detection in encoder
